Ignore the whole header when header's key is nil
On Scala client implementation when the header's key is null an
exception is thrown.

On Elixir world one tend to avoid as much as possible throwing
exceptions so this commit took the approach to just ignore the header
keeping other information of record.

The rational for that decision is that a header without a key has no
meaning and become useless, so removing it sounds reasonable