handle incomplete records when parsing record batch
This changeset relaxes the parsing of record batches when the data
returned does not align exactly with the format of a record batch. We
will return all successfully parsed record batches, and discard record
batches that are incomplete.

When specifying `max_bytes` on the fetch request, the fetch will return
exactly max_bytes. If this value is less than the available data to
fetc, the fetch request will potentially return an incomplete record
batch. In this case, we cowardly refuse to parse the record batch.

Previously we would return a match error when parsing the magic byte or
deserializing the record itself. This prevented consumers from
progressing once this situation occurred.